Fixed import paths in moved files
Several files now have correct absolute import paths following directory structure changes.
